refactor: Rename hash package to avoid conflict with function hash.
